// state/Services/services.action-types.ts
export enum ActionTypes {
  SERVICES_UPLOAD_REQUEST = 'SERVICES_UPLOAD_REQUEST',
  SERVICES_UPLOAD_SUCCESS = 'SERVICES_UPLOAD_SUCCESS',
  SERVICES_UPLOAD_FAIL = 'SERVICES_UPLOAD_FAIL',

  SERVICES_FETCH_REQUEST = 'SERVICES_FETCH_REQUEST',
  SERVICES_FETCH_SUCCESS = 'SERVICES_FETCH_SUCCESS',
  SERVICES_FETCH_FAIL = 'SERVICES_FETCH_FAIL',

  SERVICES_DELETE_REQUEST = 'SERVICES_DELETE_REQUEST',
  SERVICES_DELETE_SUCCESS = 'SERVICES_DELETE_SUCCESS',
  SERVICES_DELETE_FAIL = 'SERVICES_DELETE_FAIL',

  SERVICES_EDIT_REQUEST = 'SERVICES_EDIT_REQUEST',
  SERVICES_EDIT_SUCCESS = 'SERVICES_EDIT_SUCCESS',
  SERVICES_EDIT_FAIL = 'SERVICES_EDIT_FAIL',

  SERVICES_RESET = 'SERVICES_RESET',
}